Paralell Database Implemented
Database implemented in background which will have data written to it when the scanning is completed.
Next step is to push data to front end instead of passing the data directly. This should greatly reduce memory usage at the cost of a slight increase in scan times due to writing to disk.